Cleanup: Move around Loader components.
* Move `<LoadingOverlay>` and `<LoadingWrapper>` into common/Loaders.
* Move `<JoreLoader>` and `<Spinner>` into router/.
* Move `LoadingState` into types/ from redux/.
