+++ ui/src/components/map/terminals/Terminals.tsx
@@ -8,7 +8,6 @@ import {
 } from 'react';
 import { useAppAction, useAppSelector } from '../../../hooks';
 import {
-  LoadingState,
   MapEntityEditorViewState,
   Operation,
   isEditorOpen,
@@ -20,6 +19,7 @@ import {
   setSelectedStopIdAction,
   setSelectedTerminalIdAction,
 } from '../../../redux';
+import { LoadingState } from '../../../types';
 import { mapLngLatToGeoJSON, none } from '../../../utils';
 import { useLoader } from '../../common/hooks/useLoader';
 import { useCreateTerminal } from '../../stop-registry/terminals/hooks';
