+++ ui/src/components/routes-and-lines/search/SearchResultPage.tsx
@@ -2,10 +2,10 @@ import { FC, useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import { useLocation } from 'react-router';
 import { Path } from '../../../router/routeDetails';
-import { LoadingWrapper } from '../../../uiComponents/LoadingWrapper';
 import { PageTitle } from '../../common';
 import { CloseIconButton } from '../../common/Buttons';
 import { Container, Row, Visible } from '../../common/LayoutComponents';
+import { LoadingWrapper } from '../../common/Loaders';
 import { CompatPagination, usePagination } from '../../common/Pagination';
 import { ExportToolbar } from './ExportToolbar';
 import { FiltersContainer } from './filters/FiltersContainer';
