+++ ui/src/components/stop-registry/search/by-line/StopsByLineSearchResults.tsx
@@ -1,6 +1,6 @@
 import { FC } from 'react';
 import { useTranslation } from 'react-i18next';
-import { LoadingWrapper } from '../../../../uiComponents/LoadingWrapper';
+import { LoadingWrapper } from '../../../common/Loaders';
 import { SortStopsBy } from '../types';
 import { useStopSearchRouterState } from '../utils';
 import { StopsByLineNongroupedStopsResults } from './StopsByLineNongroupedStopsResults';
