+++ ui/src/components/stop-registry/stop-areas/versions/copy-stop-area/components/CopyStopAreaModal.tsx
@@ -4,7 +4,7 @@ import { useTranslation } from 'react-i18next';
 import { useNavigate } from 'react-router';
 import { Path, routeDetails } from '../../../../../../router/routeDetails';
 import { EnrichedStopPlace } from '../../../../../../types';
-import { LoadingWrapper } from '../../../../../../uiComponents/LoadingWrapper';
+import { LoadingWrapper } from '../../../../../common/Loaders';
 import { Modal, ModalBody, ModalHeader } from '../../../../../common/Modals';
 import { useWrapInContextNavigation } from '../../../../../forms/common/NavigationBlocker';
 import { CopyStopAreaSuccessResult } from '../../types';
