+++ ui/src/components/stop-registry/stops/stop-details/StopDetailsPage.tsx
@@ -9,8 +9,8 @@ import {
 import { Path, routeDetails } from '../../../../router/routeDetails';
 import { mapToShortDate, mapToShortDateTime } from '../../../../time';
 import { Priority } from '../../../../types/enums';
-import { LoadingWrapper } from '../../../../uiComponents/LoadingWrapper';
 import { Container, Visible } from '../../../common/LayoutComponents';
+import { LoadingWrapper } from '../../../common/Loaders';
 import { navigationBlockerContext } from '../../../forms/common/NavigationBlocker';
 import { HistoricalStopDataProvider } from '../change-history/components/HistoricalStopDataProvider';
 import { LatestStopChangeHistoryTable } from '../change-history/components/LatestStopChangeHistoryTable';
