+++ ui/src/components/stop-registry/stops/stop-details/stop-version/EditStopModal.tsx
@@ -4,7 +4,7 @@ import { useNavigate } from 'react-router';
 import { useObservationDateQueryParam } from '../../../../../hooks';
 import { Path, routeDetails } from '../../../../../router/routeDetails';
 import { StopWithDetails } from '../../../../../types';
-import { LoadingWrapper } from '../../../../../uiComponents/LoadingWrapper';
+import { LoadingWrapper } from '../../../../common/Loaders';
 import { Modal, ModalBody, ModalHeader } from '../../../../common/Modals';
 import { useWrapInContextNavigation } from '../../../../forms/common/NavigationBlocker';
 import { EditStopForm } from './EditStopForm';
