+++ ui/src/components/stop-registry/terminals/TerminalDetailsPage.tsx
@@ -2,7 +2,8 @@ import compact from 'lodash/compact';
 import { FC, useContext, useEffect, useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import { MdWarning } from 'react-icons/md';
-import { LoadingState, Operation } from '../../../redux';
+import { Operation } from '../../../redux';
+import { LoadingState } from '../../../types';
 import { useLoader } from '../../common/hooks';
 import { Container, Visible } from '../../common/LayoutComponents';
 import { navigationBlockerContext } from '../../forms/common/NavigationBlocker';
