+++ ui/src/components/timetables/substitute-day-settings/CommonSubstitutePeriod/CommonSubstitutePeriodSection.tsx
@@ -2,8 +2,8 @@ import { FC, useMemo } from 'react';
 import { useTranslation } from 'react-i18next';
 import { DateRange } from '../../../../types';
 import { SubstituteDayOfWeek } from '../../../../types/enums';
-import { LoadingWrapper } from '../../../../uiComponents/LoadingWrapper';
 import { showDangerToastWithError, showSuccessToast } from '../../../../utils';
+import { LoadingWrapper } from '../../../common/Loaders';
 import { useCreateSubstituteOperatingPeriod } from '../hooks/useCreateSubstituteOperatingPeriod';
 import { useDeleteSubstituteOperatingPeriod } from '../hooks/useDeleteSubstituteOperatingPeriod';
 import { useEditSubstituteOperatingPeriod } from '../hooks/useEditSubstituteOperatingPeriod';
