+++ ui/src/redux/slices/loader.ts
@@ -1,4 +1,5 @@
 import { PayloadAction, createSlice } from '@reduxjs/toolkit';
+import { LoadingState } from '../../types';
 
 export enum Operation {
   LoadMap = 'loadMap',
@@ -32,31 +33,6 @@ export enum Operation {
   UpdateRouteJourneyPattern = 'updateRouteJourneyPattern',
   UpdateLine = 'updateLine',
   CreateMirrorQuay = 'createMirrorQuay',
-}
-
-/**
- * Alternatively LoadingPriority.
- * Should the spinner for the {@link Operation} be hidden, or should be shown
- * based on the set priority.
- */
-export enum LoadingState {
-  /** No spinner should be visible */
-  NotLoading = 'notLoading',
-
-  /** Spinner would be a distraction to the user.
-   *  Prefer to silently load data on the background.
-   *  Only show spinner if loading takes a real long time.
-   * */
-  LowPriority = 'lowPriority',
-
-  /** Network operation should complete quickly,
-   * but if it doesn't it would confuse the user.
-   * Thus spinner should be shown soonish.
-   * */
-  MediumPriority = 'mediumPriority',
-
-  /** Block further user action immediately, until network operation is done. */
-  HighPriority = 'highPriority',
 }
 
 export function getHighestLoadingState(
