+++ ui/src/types/index.ts
@@ -3,6 +3,7 @@ export * from './DateRange';
 export * from './FilterableStopInfo';
 export * from './GeoHelpers';
 export * from './KeyHelpers';
+export * from './LoadingState';
 export * from './PagingInfo';
 export * from './SortOrder';
 export * from './StopSearchConditions';
