+++ ui/src/uiComponents/index.ts
@@ -1,7 +1,5 @@
 export * from './ConfirmationDialog';
 export * from './DialogWithButtons';
 export * from './IconToggle';
-export * from './LoadingOverlay';
 export * from './Popover';
-export * from './Spinner';
 export * from './Switch';
